requirements.txt
asttokens==2.0.8
attrs==22.1.0
autopep8==1.7.0
backcall==0.2.0
certifi==2022.6.15.1
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.5
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
entrypoints==0.4
executing==1.0.0
Fiona @ https://download.lfd.uci.edu/pythonlibs/archived/Fiona-1.8.21-cp310-cp310-win_amd64.whl
fonttools==4.37.1
GDAL @ https://download.lfd.uci.edu/pythonlibs/archived/GDAL-3.4.3-cp310-cp310-win_amd64.whl
geographiclib==1.52
geopandas==0.11.1
ipykernel==6.15.2
ipython==8.5.0
jedi==0.18.1
joblib==1.1.0
jupyter-core==4.11.1
jupyter_client==7.3.5
kiwisolver==1.4.4
matplotlib==3.5.3
matplotlib-inline==0.1.6
munch==2.5.0
nest-asyncio==1.5.5
numpy==1.23.2
packaging==21.3
pandas==1.4.4
parso==0.8.3
pickleshare==0.7.5
Pillow==9.2.0
prompt-toolkit==3.0.31
psutil==5.9.2
pure-eval==0.2.2
pycodestyle==2.9.1
Pygments==2.13.0
pyparsing==3.0.9
pyproj==3.4.0
python-dateutil==2.8.2
pytz==2022.2.1
pywin32==304
PyYAML==6.0
pyzmq==23.2.1
scikit-learn==1.1.2
scipy==1.9.1
seaborn==0.12.0
Shapely==1.8.4
six==1.16.0
stack-data==0.5.0
threadpoolctl==3.1.0
toml==0.10.2
tornado==6.2
traitlets==5.3.0
typing_extensions==4.3.0
wcwidth==0.2.5
yellowbrick==1.5
